use modal instead of confirmation dialog for stop
